+++ Changelog.md
@@ -1,4 +1,9 @@
-***Bumped version to 3.1.1-build.a1***
+***Bumped version to 3.2.0***
 
 **Added:**
-- Support for 1.21.4 for mod version 3.1.1
+- Support for Tree Harvester (and probably similar mods)
+
+**Changed:**
+- Hammering will now make mining a bit slower with Efficiency or Haste (~4%)
+- Netherite hammers are now a bit stronger
+- All hammers now have a default durability modifier of 2.5